fix: Rename account to settings everywhere
