+++ frontend/src/components/common/nav-bar/NavBar.tsx
@@ -50,7 +50,7 @@ const NavBar = () => {
         <NavLink className={styles.link} activeClassName={styles.active} exact to="/campaigns">Campaigns</NavLink>
         <a className={cx(styles.link, { [styles.active]: isCreatePath() })} onClick={handleCreateCampaign}>Create</a>
         <a className={styles.link} href={GUIDE_URL} target="_blank" rel="noopener noreferrer">Guide</a>
-        <NavLink className={styles.link} activeClassName={styles.active} to="/account">Settings</NavLink>
+        <NavLink className={styles.link} activeClassName={styles.active} to="/settings">Settings</NavLink>
 
         <div className={styles.separator}></div>
