--- frontend/src/components/dashboard/account/index.ts
@@ -1,1 +1,0 @@
-export { default } from './Account'
