+++ frontend/src/components/dashboard/settings/Settings.tsx
@@ -1,14 +1,14 @@
 import React, { useState, useEffect } from 'react'
 
 import { TitleBar } from 'components/common'
-import { UserCredential, getUserSettings } from 'services/account.service'
+import { UserCredential, getUserSettings } from 'services/settings.service'
 
 import Credentials from './credentials'
 import ApiKey from './api-key'
 
-import styles from './Account.module.scss'
+import styles from './Settings.module.scss'
 
-const Account = () => {
+const Settings = () => {
   const [hasApiKey, setHasApiKey] = useState(false)
   const [creds, setCreds] = useState([] as UserCredential[])
 
@@ -34,4 +34,4 @@ const Account = () => {
   )
 }
 
-export default Account
+export default Settings
