+++ frontend/src/components/dashboard/settings/api-key/index.ts
