+++ frontend/src/services/settings.service.ts
